Move seed phrase to the settings page
