+++ src/views/Lightning.vue
@@ -60,10 +60,9 @@
               </svg>
             </template>
             <b-dropdown-item href="#" @click="showNodeInfo">Node information</b-dropdown-item>
-            <b-dropdown-item href="#" @click="showSeed">Secret words</b-dropdown-item>
             <!-- <b-dropdown-item href="#" disabled>Check for updates</b-dropdown-item> -->
-            <b-dropdown-divider />
-            <b-dropdown-item variant="danger" href="#" disabled>Stop Lightning</b-dropdown-item>
+            <!-- <b-dropdown-divider /> -->
+            <!-- <b-dropdown-item variant="danger" href="#" disabled>Stop Lightning</b-dropdown-item> -->
           </b-dropdown>
           <b-modal id="node-info-modal" ref="node-info-modal" size="lg" centered hide-footer>
             <template v-slot:modal-header="{ close }">
@@ -118,69 +117,6 @@
                     v-else
                   ></span>
                 </div>
-              </div>
-            </div>
-          </b-modal>
-
-          <b-modal id="seed-modal" ref="seed-modal" centered hide-footer>
-            <template v-slot:modal-header="{ close }">
-              <div class="px-2 px-sm-3 pt-2 d-flex justify-content-between w-100">
-                <h3>seed phrase</h3>
-                <!-- Emulate built in modal header close button action -->
-                <a href="#" class="align-self-center" v-on:click.stop.prevent="close">
-                  <svg
-                    width="18"
-                    height="18"
-                    viewBox="0 0 18 18"
-                    fill="none"
-                    xmlns="http://www.w3.org/2000/svg"
-                  >
-                    <path
-                      fill-rule="evenodd"
-                      clip-rule="evenodd"
-                      d="M13.6003 4.44197C13.3562 4.19789 12.9605 4.19789 12.7164 4.44197L9.02116 8.1372L5.32596 4.442C5.08188 4.19792 4.68615 4.19792 4.44207 4.442C4.198 4.68607 4.198 5.0818 4.44207 5.32588L8.13728 9.02109L4.44185 12.7165C4.19777 12.9606 4.19777 13.3563 4.44185 13.6004C4.68592 13.8445 5.08165 13.8445 5.32573 13.6004L9.02116 9.90497L12.7166 13.6004C12.9607 13.8445 13.3564 13.8445 13.6005 13.6004C13.8446 13.3563 13.8446 12.9606 13.6005 12.7165L9.90505 9.02109L13.6003 5.32585C13.8444 5.08178 13.8444 4.68605 13.6003 4.44197Z"
-                      fill="#6c757d"
-                    />
-                  </svg>
-                </a>
-              </div>
-            </template>
-            <div class="px-2 px-sm-3 pb-2 pb-sm-3">
-              <!-- <span>Enter your password to view your 24-word seed phrase</span> -->
-              <div v-if="!seed.length">
-                <input-password
-                  v-model="password"
-                  ref="password"
-                  placeholder="Password"
-                  inputGroupClass="neu-input-group"
-                  :inputClass="[
-                    isIncorrectPassword ? 'incorrect-password' : '',
-                    'form-control form-control-lg neu-input w-100'
-                  ]"
-                  :disabled="isLoadingSeed"
-                />
-
-                <small
-                  class="mt-2 text-danger error float-right"
-                  v-show="isIncorrectPassword"
-                >Incorrect password</small>
-
-                <b-button
-                  variant="success"
-                  class="mt-3 mb-2"
-                  :disabled="password && isLoadingSeed"
-                  @click="fetchSeed"
-                >
-                  {{
-                  isLoadingSeed ? "Decrypting Seed..." : "View Seed"
-                  }}
-                </b-button>
-              </div>
-
-              <div class="d-flex justify-content-center" v-else>
-                <!-- Seed phrase -->
-                <seed :words="seed" v-if="seed.length"></seed>
-                <b-spinner v-else></b-spinner>
               </div>
             </div>
           </b-modal>
@@ -328,8 +264,6 @@ import Stat from "@/components/Utility/Stat";
 import LightningWallet from "@/components/LightningWallet";
 import QrCode from "@/components/Utility/QrCode";
 import InputCopy from "@/components/Utility/InputCopy";
-import InputPassword from "@/components/Utility/InputPassword";
-import Seed from "@/components/Utility/Seed";
 import ChannelList from "@/components/Channels/List";
 import ChannelOpen from "@/components/Channels/Open";
 import ChannelManage from "@/components/Channels/Manage";
@@ -338,10 +272,7 @@ export default {
   data() {
     return {
       status: "Running",
-      selectedChannel: {},
-      password: "",
-      isIncorrectPassword: false,
-      isLoadingSeed: false
+      selectedChannel: {}
     };
   },
   computed: {
@@ -355,31 +286,12 @@ export default {
       pubkey: state => state.lightning.pubkey,
       uris: state => state.lightning.uris,
       channels: state => state.lightning.channels,
-      unit: state => state.system.unit,
-      seed: state => state.user.seed
+      unit: state => state.system.unit
     })
   },
   methods: {
     showNodeInfo() {
       this.$refs["node-info-modal"].show();
-    },
-    async fetchSeed() {
-      this.isLoadingSeed = true;
-      try {
-        await this.$store.dispatch("user/getSeed", this.password);
-      } catch (error) {
-        if (
-          error.response &&
-          error.response.data &&
-          error.response.data === "Incorrect password"
-        ) {
-          this.isIncorrectPassword = true;
-        }
-      }
-      this.isLoadingSeed = false;
-    },
-    showSeed() {
-      this.$refs["seed-modal"].show();
     },
     async downloadChannelBackup() {
       await API.download(
@@ -435,8 +347,6 @@ export default {
     Stat,
     QrCode,
     InputCopy,
-    InputPassword,
-    Seed,
     ChannelList,
     ChannelOpen,
     ChannelManage
