+++ src/views/Settings.vue
@@ -1,9 +1,68 @@
 <template>
   <div class="p-sm-2">
     <div class="my-3 pb-2">
-      <h1>settings</h1>
-      <!-- <p class="text-muted">Your umbrel, your rules.</p> -->
+      <div class="d-flex justify-content-between align-items-center">
+        <h1>settings</h1>
+        <b-dropdown variant="link" toggle-class="text-decoration-none p-0" no-caret right>
+          <template v-slot:button-content>
+            <svg
+              width="18"
+              height="4"
+              viewBox="0 0 18 4"
+              fill="none"
+              xmlns="http://www.w3.org/2000/svg"
+            >
+              <path
+                fill-rule="evenodd"
+                clip-rule="evenodd"
+                d="M2 4C3.10457 4 4 3.10457 4 2C4 0.89543 3.10457 0 2 0C0.89543 0 0 0.89543 0 2C0 3.10457 0.89543 4 2 4Z"
+                fill="#C3C6D1"
+              />
+              <path
+                fill-rule="evenodd"
+                clip-rule="evenodd"
+                d="M9 4C10.1046 4 11 3.10457 11 2C11 0.89543 10.1046 0 9 0C7.89543 0 7 0.89543 7 2C7 3.10457 7.89543 4 9 4Z"
+                fill="#C3C6D1"
+              />
+              <path
+                fill-rule="evenodd"
+                clip-rule="evenodd"
+                d="M16 4C17.1046 4 18 3.10457 18 2C18 0.89543 17.1046 0 16 0C14.8954 0 14 0.89543 14 2C14 3.10457 14.8954 4 16 4Z"
+                fill="#C3C6D1"
+              />
+            </svg>
+          </template>
+          <b-dropdown-item href="#" v-b-modal.seed-modal>View secret words</b-dropdown-item>
+        </b-dropdown>
+      </div>
     </div>
+
+    <b-modal id="seed-modal" centered hide-footer>
+      <template v-slot:modal-header="{ close }">
+        <div class="px-2 px-sm-3 pt-2 d-flex justify-content-between w-100">
+          <h3>secret words</h3>
+          <!-- Emulate built in modal header close button action -->
+          <a href="#" class="align-self-center" v-on:click.stop.prevent="close">
+            <svg
+              width="18"
+              height="18"
+              viewBox="0 0 18 18"
+              fill="none"
+              xmlns="http://www.w3.org/2000/svg"
+            >
+              <path
+                fill-rule="evenodd"
+                clip-rule="evenodd"
+                d="M13.6003 4.44197C13.3562 4.19789 12.9605 4.19789 12.7164 4.44197L9.02116 8.1372L5.32596 4.442C5.08188 4.19792 4.68615 4.19792 4.44207 4.442C4.198 4.68607 4.198 5.0818 4.44207 5.32588L8.13728 9.02109L4.44185 12.7165C4.19777 12.9606 4.19777 13.3563 4.44185 13.6004C4.68592 13.8445 5.08165 13.8445 5.32573 13.6004L9.02116 9.90497L12.7166 13.6004C12.9607 13.8445 13.3564 13.8445 13.6005 13.6004C13.8446 13.3563 13.8446 12.9606 13.6005 12.7165L9.90505 9.02109L13.6003 5.32585C13.8444 5.08178 13.8444 4.68605 13.6003 4.44197Z"
+                fill="#6c757d"
+              />
+            </svg>
+          </a>
+        </div>
+      </template>
+      <seed></seed>
+    </b-modal>
+
     <b-row>
       <b-col col cols="12" md="6" xl="4">
         <card-widget
@@ -196,6 +255,7 @@ import API from "@/helpers/api";
 
 import CardWidget from "@/components/CardWidget";
 import ToggleSwitch from "@/components/ToggleSwitch";
+import Seed from "@/components/Seed";
 import InputPassword from "@/components/Utility/InputPassword";
 import InputCopy from "@/components/Utility/InputCopy";
 
@@ -413,7 +473,8 @@ export default {
     CardWidget,
     ToggleSwitch,
     InputPassword,
-    InputCopy
+    InputCopy,
+    Seed
   }
 };
 </script>
